refactor: move shared catalog view builders into @repo/shared (#1514)
