+++ app/views/EntityView/assembly/components/Side/ga2/side.tsx
@@ -1,11 +1,11 @@
-import {
-  buildAssemblyResources,
-  buildOrganismDetails,
-} from "@/viewModelBuilders/catalog/common/viewModelBuilders";
 import { AssemblyDetails } from "@/views/EntityView/assembly/components/Side/ga2/components/AssemblyDetails/AssemblyDetails";
 import { mapAssemblyToOrganism } from "@/views/WorkflowInputsView/utils";
 import { BackPageContentSideColumn } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/backPageView.styles";
 import { AssemblyFavoriteButton } from "@repo/shared/components/Favorites/components/AssemblyFavoriteButton/assemblyFavoriteButton";
+import {
+  buildAssemblyResources,
+  buildOrganismDetails,
+} from "@repo/shared/viewModelBuilders/viewModelBuilders";
 import { AnalysisPortals } from "@repo/shared/views/EntityView/assembly/components/Side/AnalysisPortals/analysisPortals";
 import { KeyValueSection } from "@repo/shared/views/EntityView/components/KeyValueSection/keyValueSection";
 import { JSX } from "react";
